+++ src/app/app-routing.module.ts
@@ -5,7 +5,6 @@ import { PreloadAllModules } from '@angular/router';
 import { HomePageComponent } from './home-page/home-page.component';
 import { LoadingComponent } from './loading/loading.component';
 import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
-import { TestComponent } from './test/test.component';
 
 const routes: Routes = [
   {
@@ -23,10 +22,6 @@ const routes: Routes = [
   {
     path: "loading",
     component: LoadingComponent,
-  },
-  {
-    path: "test",
-    component: TestComponent,
   },
   {
     path: '**', component: PageNotFoundComponent