+++ src/app/highlight-search.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { HighlightSearchPipe } from './highlight-search.pipe';
+
+describe('HighlightSearchPipePipe', () => {
+  it('create an instance', () => {
+    const pipe = new HighlightSearchPipe();
+    expect(pipe).toBeTruthy();
+  });
+});